Share the sql_store fixture through a conftest
