+++ tests/unit/zen_stores/conftest.py
@@ -1,0 +1,38 @@
+#  Copyright (c) ZenML GmbH 2026. All Rights Reserved.
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at:
+#
+#       https://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+#  or implied. See the License for the specific language governing
+#  permissions and limitations under the License.
+"""Fixtures shared by the store unit tests."""
+
+from pathlib import Path
+from typing import Iterator
+
+import pytest
+
+from zenml.zen_stores.sql_zen_store import (
+    SqlZenStore,
+    SqlZenStoreConfiguration,
+)
+
+
+@pytest.fixture
+def sql_store(
+    tmp_path: Path, monkeypatch: pytest.MonkeyPatch
+) -> Iterator[SqlZenStore]:
+    """Create a fresh SQLite-backed SqlZenStore for tests."""
+    db_dir = tmp_path / "zenml-cfg"
+    db_dir.mkdir(parents=True, exist_ok=True)
+    monkeypatch.setenv("ZENML_CONFIG_PATH", str(db_dir))
+    db_path = db_dir / "test.db"
+    config = SqlZenStoreConfiguration(url=f"sqlite:///{db_path}")
+    store = SqlZenStore(config=config, skip_default_registrations=False)
+    yield store
