+++ tests/unit/zen_stores/test_run_deletion_cleanup.py
@@ -36,28 +36,6 @@ from zenml.zen_stores.sql_zen_store import (
 )
 
 
-@pytest.fixture
-def sql_store(
-    tmp_path: Path, monkeypatch: pytest.MonkeyPatch
-) -> Iterator[SqlZenStore]:
-    """Create a fresh SQLite-backed SqlZenStore for tests.
-
-    Args:
-        tmp_path: Temporary directory for the database file.
-        monkeypatch: Fixture used to point the ZenML config path at it.
-
-    Yields:
-        The store.
-    """
-    db_dir = tmp_path / "zenml-cfg"
-    db_dir.mkdir(parents=True, exist_ok=True)
-    monkeypatch.setenv("ZENML_CONFIG_PATH", str(db_dir))
-    db_path = db_dir / "test.db"
-    config = SqlZenStoreConfiguration(url=f"sqlite:///{db_path}")
-    store = SqlZenStore(config=config, skip_default_registrations=False)
-    yield store
-
-
 def _run_with_step_metadata(store: SqlZenStore) -> Tuple[UUID, UUID]:
     """Create a run with one step run that has metadata attached to it.
